backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
celery==5.3.4
python-docx==1.1.0
pdfminer.six==20231228
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.25.2
faiss-cpu==1.7.4
openai==1.3.8
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1